@@ -1,16 +1,16 @@
 """Perlin Noise calculating lib."""
+
+import itertools
 import math
 import random
-from collections.abc import Iterable
-from typing import Dict, Optional, Tuple, Union, List
-import itertools
 from functools import lru_cache
+from typing import Dict, List, Optional, Tuple, Union
 
 from perlin_noise.rand_vec import RandVec
 from perlin_noise.tools import hasher
 
 
-class PerlinNoise(object):
+class PerlinNoise:
     """Smooth random noise generator.
 
     read more https://en.wikipedia.org/wiki/Perlin_noise
@@ -31,16 +31,20 @@ class PerlinNoise(object):
             ValueError: if seed is negative
         """
         if octaves <= 0:
-            raise ValueError('octaves expected to be positive number')
+            raise ValueError("octaves expected to be positive number")
 
         if seed is not None and not isinstance(seed, int) and seed <= 0:
-            raise ValueError('seed expected to be positive integer number')
+            raise ValueError("seed expected to be positive integer number")
 
         self.octaves: float = octaves
         self.seed: int = seed if seed else random.randint(1, 10**5)  # noqa: S311, E501
         self.cache: Dict[Tuple, RandVec] = {}
 
-    def __call__(self, coordinates: Union[int, float, Iterable], tile_sizes: Optional[Union[int, List, Tuple]] = None) -> float:
+    def __call__(
+        self,
+        coordinates: Union[int, float, List, Tuple],
+        tile_sizes: Optional[Union[int, List, Tuple]] = None,
+    ) -> float:
         """Forward request to noise function.
 
         Parameters:
@@ -52,7 +56,11 @@ class PerlinNoise(object):
         """
         return self.noise(coordinates, tile_sizes)
 
-    def noise(self, coordinates: Union[int, float, Iterable], tile_sizes: Optional[Union[int, List, Tuple]] = None) -> float:
+    def noise(  # noqa: WPS231
+        self,
+        coordinates: Union[int, float, List, Tuple],
+        tile_sizes: Optional[Union[int, List, Tuple]] = None,
+    ) -> float:
         """Get perlin noise value for given coordinates.
 
         Parameters:
@@ -64,47 +72,55 @@ class PerlinNoise(object):
 
         Raises:
             TypeError: if coordinates is not valid type
+            ValueError: if tile_sizes have different length than coordinates
         """
-        if not isinstance(coordinates, (int, float, Iterable)):
-            raise TypeError('coordinates must be int, float or iterable')
+        if not isinstance(coordinates, (int, float, list, tuple)):
+            raise TypeError("coordinates must be int, float or iterable")
 
-        if tile_sizes is not None:
-            if not (isinstance(tile_sizes, int) or (isinstance(tile_sizes, (tuple, list)) and all(isinstance(tile, int) for tile in tile_sizes))):
-                raise TypeError('tile_sizes must be int or list of int')
-
-        
         if isinstance(coordinates, (int, float)):
             coordinates = [coordinates]
         else:
             coordinates = list(coordinates)
 
-    
         if tile_sizes is not None:
             if isinstance(tile_sizes, int):
                 tile_sizes = [tile_sizes]
+            # fmt: off
+            if not all(isinstance(tile, int) for tile in tile_sizes):
+                raise TypeError("tile_sizes must be int or list of int")
+            # fmt: on
+
             if len(tile_sizes) != len(coordinates):
-                raise ValueError('tile_sizes must have same length as coordinates')
-            
-            coordinates = [coordinate % tile for coordinate, tile in zip(coordinates, tile_sizes)]
+                raise ValueError("tile_sizes must have same length as coordinates")
+
+            coordinates = [coors % tile for coors, tile in zip(coordinates, tile_sizes)]
             tile_sizes = tuple(tile * self.octaves for tile in tile_sizes)
         coordinates = [coordinate * self.octaves for coordinate in coordinates]
-    
 
         coor_bounding_box = [
             (math.floor(coordinate), math.floor(coordinate) + 1)
             for coordinate in coordinates
         ]
-        return sum([
-            self.get_from_cache_of_create_new(coors, tile_sizes).get_weighted_val(coordinates)
-            for coors in itertools.product(*coor_bounding_box)
-        ])
+        return sum(
+            [
+                self.get_from_cache_of_create_new(coors, tile_sizes).get_weighted_val(
+                    coordinates,
+                )
+                for coors in itertools.product(*coor_bounding_box)
+            ]  # noqa: C812
+        )
 
-    @lru_cache(maxsize=100, typed=False)
-    def get_from_cache_of_create_new(self, coors: Tuple[int, ...], tile_sizes: Optional[Tuple[int, ...]] = None) -> RandVec:
+    @lru_cache(maxsize=100, typed=False)  # noqa: B019
+    def get_from_cache_of_create_new(
+        self,
+        coors: Tuple[int, ...],
+        tile_sizes: Optional[Tuple[int, ...]] = None,
+    ) -> RandVec:
         """Use cached RandVec or creates new.
 
         Parameters:
             coors: Tuple of int vector coordinates
+            tile_sizes: optional tile sizes to repetative patterns
 
         Returns:
             RandVec
