@@ -1,14 +1,16 @@
 """Random vector generator used for weighting perlin noise."""
 
+from functools import reduce
+from operator import mul
 from typing import List, Tuple
 
-from perlin_noise.tools import dot, fade, product, sample_vector
+from perlin_noise.tools import dot, fade, sample_vector
 
 
-class RandVec(object):
+class RandVec:
     """Vectors to give weights and contribute in final value."""
 
-    def __init__(self, coordinates: Tuple[int], seed: int):
+    def __init__(self, coordinates: Tuple[int, ...], seed: int):
         """Vector initializer in specified coordinates.
 
         Parameters:
@@ -29,8 +31,7 @@ class RandVec(object):
 
         """
         return tuple(
-            coor1 - coor2
-            for coor1, coor2 in zip(coordinates, self.coordinates)
+            coor1 - coor2 for coor1, coor2 in zip(coordinates, self.coordinates)
         )
 
     def weight_to(self, coordinates: List[float]) -> float:
@@ -42,15 +43,12 @@ class RandVec(object):
         Returns:
             weight
         """
-        weighted_dists = list(
-            map(
-                lambda dist: fade(1 - abs(dist)),
-                self.dists_to(coordinates),
-            ))
+        # fmt: off
+        weighted_dists = [fade(1 - abs(dist)) for dist in self.dists_to(coordinates)]  # noqa: E501, WPS221
+        # fmt: on
+        return reduce(mul, weighted_dists)
 
-        return product(weighted_dists)
-
-    def get_weighted_val(self, coordinates: Tuple[float]) -> float:
+    def get_weighted_val(self, coordinates: List[float]) -> float:
         """Calculate weighted contribution of this vec to final result.
 
         Parameters:
@@ -60,5 +58,6 @@ class RandVec(object):
             weighted contribution
         """
         return self.weight_to(coordinates) * dot(
-            self.vec, self.dists_to(coordinates),
+            self.vec,
+            self.dists_to(coordinates),
         )
