@@ -3,5 +3,5 @@
 from perlin_noise.perlin_noise import PerlinNoise
 
 __all__ = [  # noqa: WPS410
-    'PerlinNoise',
+    "PerlinNoise",
 ]
