@@ -1,5 +1,6 @@
-from perlin_noise import PerlinNoise
 import random
+
+from perlin_noise import PerlinNoise
 
 
 def test_perlin_noise_works_in_1D():
@@ -8,10 +9,11 @@ def test_perlin_noise_works_in_1D():
     lim = 2000
     prev = 0
     for i in range(lim):
-        noise = p_noise([i/lim])
+        noise = p_noise([i / lim])
         assert -1 < noise < 1
         assert 0 <= abs(noise - prev) < 0.001
         prev = noise
+
 
 def test_perlin_noise_works_in_2D():
     """Check value in [-1, 1] range."""
@@ -19,7 +21,7 @@ def test_perlin_noise_works_in_2D():
     lim = 100
     for i in range(lim):
         for j in range(lim):
-            assert -1 <= noise([i/lim, j/lim]) <= 1
+            assert -1 <= noise([i / lim, j / lim]) <= 1
 
 
 def test_perlin_noise_works_in_high_dimensions():
@@ -44,12 +46,12 @@ def test_noise_value_doesnt_depend_on_space_size():
 
 
 def test_that_dist_in_0_1_range():
-    noise = PerlinNoise(octaves = 20.496, seed = 1337)
-    assert noise([125/1281, 1152/1281])
+    noise = PerlinNoise(octaves=20.496, seed=1337)
+    assert noise([125 / 1281, 1152 / 1281])
 
 
 def test_tiles_works():
-    noise = PerlinNoise(octaves = 1.496)
+    noise = PerlinNoise(octaves=1.496)
     assert noise(0.5, tile_sizes=1) == noise(0.5)
     assert noise([0.5, 0.5], tile_sizes=[43, 2345]) == noise([0.5, 0.5])
     assert noise([1.5, 1.5], tile_sizes=[1, 1]) == noise([0.5, 0.5])
@@ -57,15 +59,14 @@ def test_tiles_works():
 
 
 def test_tiles_seamless():
-    p_noise = PerlinNoise(octaves = 2)
+    p_noise = PerlinNoise(octaves=2)
     lim = 400
     prev = 0
     x_lim, y_lim = 5, 10
     for i in range(0, lim, 100):
-        prev = p_noise([x_lim * i/lim, 0], tile_sizes=(2, 3))
+        prev = p_noise([x_lim * i / lim, 0], tile_sizes=[2, 3])
         for j in range(lim):
-            noise = p_noise([x_lim * i/lim, y_lim * j/lim], tile_sizes=(2, 3))
+            noise = p_noise([x_lim * i / lim, y_lim * j / lim], tile_sizes=[2, 3])
             assert -1 < noise < 1
             assert 0 <= abs(noise - prev) < 0.1
             prev = noise
-
